backend/pyproject.toml
[tool.poetry]
name = "past-years"
version = "0.1.0"
description = ""
authors = ["Guacs <guacs.guacs@gmail.com>"]
readme = "README.md"
packages = [{ include = "past_years" }]

[tool.poetry.dependencies]
python = "^3.11"
msgspec = "^0.13.1"
loguru = "^0.6.0"
whoosh = "^2.7.4"
falcon = "^3.1.1"
python-dotenv = "^1.0.0"
httpx = "^0.23.3"
mysql-connector-python = "^8.0.32"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.1.0"
ruff = "^0.0.252"
black = "^23.1.0"
click = "^8.1.3"
pytest = "^7.2.1"

[tool.ruff]
fix = true
select = [
    # isort
    "I001",

    # remove unused imports
    "F401",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
